backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Pillow==10.4.0
djoser==2.2.3
drf-spectacular==0.27.2
django-filter==24.3
django-debug-toolbar==4.4.6
redis==5.2.1
redis[hiredis]==5.2.1